Manually write a packer for width of 1.
This is written in a way that avoid writing and reading a single
register for every single input value, instead choosing to combine
partial values in a tree structure to compute the final value.

This is slightly less negative on my local host but also substantially
changes the way the loop would run, it would be interesting to
benchmark this on scalar on its own. Notably, this is still slower than
the scalar baseline which we did not recompute (even though it would be
affected by this change).

OLD
BitPacker4x/compress-1  time:   [140.35 ns 140.60 ns 140.86 ns]
                        thrpt:  [9.0873 Gelem/s 9.1037 Gelem/s 9.1199 Gelem/s]
                 change:
                        time:   [+4.9019% +5.2543% +5.5877%] (p = 0.00 < 0.05)
                        thrpt:  [-5.2920% -4.9920% -4.6729%]
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

NEW
BitPacker4x/compress-1  time:   [138.57 ns 138.79 ns 139.05 ns]
                        thrpt:  [9.2056 Gelem/s 9.2223 Gelem/s 9.2373 Gelem/s]
                 change:
                        time:   [+3.8068% +4.1283% +4.4440%] (p = 0.00 < 0.05)
                        thrpt:  [-4.2549% -3.9646% -3.6672%]
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high mild